uploadpost 에서 uploadTag를 컴포넌트로 분리
